+++ app/controllers/api_controller.rb
@@ -38,7 +38,17 @@ class ApiController < ApplicationController
   end
 
   def authenticate_active_api_user!
-    render json: { error: 'User is not active' }, status: :unauthorized unless current_api_user&.active_until&.future?
+    if current_api_user.nil?
+      render json: { error: 'User account is not active or has been deleted' }, status: :unauthorized
+
+      return false
+    end
+
+    if current_api_user.active_until&.past?
+      render json: { error: 'User subscription is not active' }, status: :unauthorized
+
+      return false
+    end
 
     true
   end