+++ app/jobs/data_migrations/migrate_points_latlon_job.rb
@@ -4,7 +4,7 @@ class DataMigrations::MigratePointsLatlonJob < ApplicationJob
   queue_as :data_migrations
 
   def perform(user_id)
-    user = User.find(user_id)
+    user = find_user_or_skip(user_id) || return
 
     user.points.update_all('lonlat = ST_SetSRID(ST_MakePoint(longitude, latitude), 4326)')
     # rubocop:enable Rails/SkipsModelValidations