+++ app/jobs/users/export_data_job.rb
@@ -6,7 +6,7 @@ class Users::ExportDataJob < ApplicationJob
   sidekiq_options retry: false
 
   def perform(user_id)
-    user = User.find(user_id)
+    user = find_user_or_skip(user_id) || return
 
     Users::ExportData.new(user).export
   end