+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.1] - 2026-02-27
+
+### Changed
+
+- User deletion now being done in the background to prevent request timeouts for users with large amount of data.
+
+### Fixed
+
+- Point speed in Map V2 is now correctly calculated from m/s to km/h or mph based on user preference. #2308
+- Family members are now being loaded correctly on Map V2 when family layer is enabled. #2250
+- Photos popups on Map V2 now show the photo timestamp in user's timezone. #2310
+- Fix the issue preventing fresh app from starting. #2304
+
 ## [1.3.0] - 2026-02-25
 
 The Storage & Timeline Interaction Release